changed type of refreshinterval to long

+++ bundles/binding/org.openhab.binding.http/src/main/java/org/openhab/binding/http/internal/HttpInBinding.java
@@ -148,7 +148,7 @@ public class HttpInBinding extends AbstractActiveBinding<HttpBindingProvider> im
      * @{inheritDoc}
      */
     @Override
-    protected int getRefreshInterval() {
+    protected long getRefreshInterval() {
     	return HttpInBinding.granularity;
     }
